requirements.txt
Flask
Flask-RESTful
Werkzeug
Flask-WTF
Flask-SQLAlchemy
Flask-CORS
sphinx
PyPDF2
WTForms
googletrans
gTTS
pytube
openai-whisper
openai
torch
pytest
flask-sqlalchemy 
flask-jwt-extended 
Flask-Migrate
pyotp 
bleach 
python-dotenv
yfinance
cachetools
phonenumbers
Flask-Limiter
email-validator
Flask-Talisman